==> 2. preprocessing with pandas.ipynb <==
저장해둔 parquet파일을 읽어온다.
용이한 전처리를 위해 판다스 데이터프레임으로 불러왔다.

import pandas as pd

df = pd.read_parquet('data/market.parquet')

df.head()

Preprocessing

1. category_code컬럼 값 정리('.'기준으로 문자열 나눠 저장)
2. category_id컬럼 삭제
3. event_time컬럼의 datetime 순으로 데이터프레임 정렬

#category_code의 맨 뒤 문자열을 추출하여 product 컬럼으로 추가
df.insert(5,'product', df['category_code'].apply(lambda x : x.split('.')[-1]))

#category_code의 중간 문자열을 추출하여 sub_category 컬럼으로 추가(없는 경우 product값으로 대체됨)
df.insert(5,'sub_category', df['category_code'].apply(lambda x : x.split('.')[1]))

#category_code 값에서 제품명을 제거, '.'대신 '/'으로 교체
df['category_code'] = df['category_code'].apply(lambda x : x.split('.')[0])

#category_id컬럼 삭제
df = df.drop('category_id', axis=1)

#event_time컬럼의 datetime 순으로 데이터프레임 정렬
df.sort_values(by=['event_time'], inplace=True)
df.reset_index(drop=True, inplace=True)

df.head()

#데이터셋 개요 확인
df.info()

#parquet파일 덮어씌워 저장
df.to_parquet('data/market_final.parquet', compression='gzip')
