==> 3. session_by_date.ipynb <==
SQL로 정제하기 어려운 데이터 처리의 경우 파이썬 판다스를 활용하기로 함

1. 날짜별 세션 로그 수와 장바구니, 구매 건수

import pandas as pd

df = pd.read_parquet('data/market_final.parquet')

#event_time 컬럼의 날짜별로 세션 로그 수 집계
session_by_date = df.groupby(df['event_time'].dt.date)['user_session'].count().to_frame()

#event_time 컬럼의 날짜별로 장바구니 건수 집계
session_by_date['cart_cnt'] = df[df['event_type'] == 'cart'].groupby(df['event_time'].dt.date)['event_type'].count()

#event_time 컬럼의 날짜별로 구매 건수 집계
session_by_date['purchase_cnt'] = df[df['event_type'] == 'purchase'].groupby(df['event_time'].dt.date)['event_type'].count()

#인덱스 정리
session_by_date.reset_index(inplace=True)

session_by_date.info()

2. 요일별 세션 로그 수와 장바구니, 구매 건수

session_by_date['event_time'] = pd.to_datetime(session_by_date['event_time'])
session_by_date.insert(1,'weekday',session_by_date['event_time'].dt.day_name())

session_by_date

#결측치 제거 및 데이터타입 정리
session_by_date = session_by_date.fillna(0)
session_by_date = session_by_date.astype({'purchase_cnt':'int64'})

#parquet 파일로 저장
session_by_date.to_parquet('data/session_by_date.parquet', compression='gzip')

이 후 빅쿼리로 옮겨 시각화 및 대쉬보드 제작 진행
