==> 5. RFM_analysis.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from sklearn import preprocessing
import seaborn as sns
from matplotlib import pyplot as plt

df = pd.read_parquet('data/market_final.parquet')

df_purchase = df[df['event_type'] == 'purchase']
df_purchase.head()

def mean(s):
    return np.mean(s)

df_monetary = df_purchase.groupby('user_id').agg({'price':'sum'})
df_monetary = df_monetary.rename(columns={'price':'monetary'}).reset_index()
df_monetary.head()

temp_recency_df = df_purchase[['user_id','event_time']].drop_duplicates()
df_recency = temp_recency_df.groupby('user_id')['event_time'].max().reset_index()
df_recency = df_recency.rename(columns={'event_time':'recency'})
df_recency.head()

temp_frequency_df = df_purchase[['user_id','event_time']].drop_duplicates()
df_frequency = temp_frequency_df.groupby('user_id')['event_time'].count().reset_index()
df_frequency = df_frequency.rename(columns={'event_time':'frequency'})
df_frequency.head()

df_rfm = pd.merge(df_recency, df_frequency, how='left', on='user_id')
df_rfm = pd.merge(df_rfm, df_monetary, how='left', on='user_id')

df_rfm

current_day = pd.to_datetime('20191101')
time_diff = df_rfm['recency'] - current_day
time_in_seconds = [x.total_seconds() for x in time_diff]
df_rfm['recency'] = time_in_seconds

df_rfm

def get_score(level, data):
    '''level 안의 원소들을 기준으로 기준을 나누어 data의 점수를 부여
    리스트 형태로 반환'''
    score = []
    
    for j in range(len(data)):
        for i in range(len(level)):
            if data[j] <= level[i]:
                score.append(i+1)
                break
            elif data[j] > max(level):
                score.append(len(level)+1)
                break
            else:
                continue
    return score

def get_rfm_grade(df, num_class, rfm_tick_point, rfm_col_map, suffix=None):
    
    # 예외처리
    if not isinstance(df, pd.DataFrame):
        print('df must be pandas.DataFrame')
        return
    
    if isinstance(rfm_tick_point, dict) == False or isinstance(rfm_col_map, dict) == False:
        print('rfm_tick_point & rfm_col_map must be dict')
        return
    
    if len(rfm_col_map) != 3:
        print(f'rfm_col_map must have three keys. \n now : {len(rfm_col_map)}')
        return
    
    if len(rfm_tick_point) != 3:
        print(f'rfm_tick_point must have three keys. \n now : {len(rfm_tick_point)}')
        return
        
    if set(rfm_tick_point.keys()) != set(rfm_col_map.keys()):
        print('rfm_tick_point and rfm_col_map must have the same key')
        return
    
    if not set(rfm_col_map.values()).issubset(set(df.columns)):
        print(f'{set(rfm_col_map.values())-set(df.columns)} must be in the columns')
        return
    
    for k, v in rfm_tick_point.items():
        if isinstance(v, str):
            if not v in ['quantile', 'min_max']:
                print(f'{k} must be either "quantile" or "min_max"')
                return
        elif isinstance(v, list) or isinstance(v, tuple):
            if len(v) != num_class-1:
                print(f'There must be {num_class-1} elements in the list (tuples) corresponding to {k}.')
                return
            
    if suffix:
        if not isinstance(suffix, str):
            print('the suffix must be "str"')
            return
    
    # 점수 부여    
    for k, v in rfm_tick_point.items():
        if isinstance(v, str):
            if v == 'quantile':
                scale = preprocessing.StandardScaler() # 데이터 표준화 함수 지정
                temp_data = np.array(df[rfm_col_map[k]]) # numpy배열로 변환
                temp_data = temp_data.reshape((-1,1)) # 스케일링을 위해 1차원 배열 -> 2차원으로 변환
                temp_data = scale.fit_transform(temp_data) # 데이터 평균 0, 표준편차 1 갖도록 스케일링
                temp_data = temp_data.squeeze() # 데이터를 다시 1차원으로 변환
                
                quantiles_level = np.linspace(0,1,num_class+1)[1:-1] # 0과 1 사이 분위수를 num_class만큼 생성
                quantiles = [] # 분위수를 담을 리스트
                for x in quantiles_level:
                    quantiles.append(np.quantile(temp_data, x)) # 분위수 리스트에 삽입
            else: # 최대 최소값 경우
                temp_data = np.array(df[rfm_col_map[k]]) # numpy배열로 변환
                quantiles = np.linspace(np.min(temp_data), np.max(temp_data), num_class+1)[1:-1] # 최대 최소값 사이 등간격으로 분할
        
        # 직접 구분값을 넣어주는 경우
        else:
            temp_data = np.array([df[rfm_col_map[k]]]) # numpy배열로 변환
            quantiles = v # 직접 구분값 넣어주기
        
        score = get_score(quantiles, temp_data) # 앞서 정의한 quantiles를 기준으로 한 점수리스트 반환
        new_col_name = rfm_col_map[k]+'_'+k # 점수값을 담는 변수의 이름
        if suffix:
            new_col_name = rfm_col_map[k]+'_'+suffix
        df[new_col_name] = score
    return df
            
            

rfm_tick_point={'R':'min_max','F':'min_max','M':'min_max'}
rfm_col_map={'R':'recency','F':'frequency','M':'monetary'}

df_result = get_rfm_grade(df=df_rfm, num_class=5, rfm_tick_point=rfm_tick_point, rfm_col_map=rfm_col_map)

df_result[['user_id','recency_R','frequency_F','monetary_M']]

print(df_result['recency_R'].value_counts(), 
      df_result['frequency_F'].value_counts(),
      df_result['monetary_M'].value_counts(), sep='\n\n')

# f, axes = plt.subplots(2, 2, figsize=(8, 6), sharex=True)
# sns.distplot(df_result['recency_R'], color='blue', label='recency_R', ax=axes[0, 0])
# sns.distplot(df_result['frequency_F'], color='red', label='frequency_F', ax=axes[0, 1])
# sns.distplot(df_result['monetary_M'], color='green', label='monetary_M', ax=axes[1, 0])
# plt.legend(title="RFM")
# plt.show()
