==> 6. recsys.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.decomposition import PCA
import annoy
import scipy.sparse as sparse
import random
from implicit.als import AlternatingLeastSquares as ALS
import implicit.evaluation

df = pd.read_parquet('data/market_final.parquet')

df.head()

제품 카테고리, 브랜드, 가격 정보를 활용한 컨텐츠 기반 필터링 모델

df_cb = df[['product_id', 'category_code', 'sub_category', 'product', 'brand', 'price']]
df_cb = df_cb.drop_duplicates(['product_id'], keep = 'first')
df_cb = df_cb.sort_values(by='product_id', ascending=True)
df_cb.reset_index(drop=True, inplace=True)

df_cb.shape

df_cb.head()

df_cb_price = df_cb['price'].to_frame()
df_cb_price.head()

# 최대 최소 범위에서 스케일링
scaler = MinMaxScaler()
price_feature_scaled = scaler.fit_transform(df_cb_price)
price_feature_scaled = pd.DataFrame(price_feature_scaled, columns=df_cb_price.columns, index=list(df_cb_price.index.values))
price_feature_scaled.head()

df_cb['product_features'] = df_cb['category_code'] + ' ' + df_cb['sub_category'] + ' ' + df_cb['product'] + ' ' + df_cb['brand']
df_cb.head()

product_lookup = df_cb['product_id'].to_frame()
product_lookup

tfidf = TfidfVectorizer()
df_tfidf = tfidf.fit_transform(df_cb['product_features'])
df_tfidf = pd.DataFrame(df_tfidf.todense(), columns = tfidf.get_feature_names_out())
df_tfidf.head()

df_tfidf.insert(0, 'product_price', price_feature_scaled)

df_tfidf

pca = PCA(n_components=100)
df_tfidf_pca = pca.fit_transform(df_tfidf)
df_tfidf_pca.shape

annoy_index = annoy.AnnoyIndex(f=df_tfidf_pca.shape[1], metric='angular')

for idx in range(df_tfidf_pca.shape[0]):
    vector = df_tfidf_pca[idx, :]
    annoy_index.add_item(idx, vector)

annoy_index.build(n_trees=10)
annoy_index.save('cb.ann')

get_nns_list = annoy_index.get_nns_by_vector(vector=df_tfidf_pca[36235], n=16, include_distances=True)
get_nns_list

df_cb.iloc[get_nns_list[0]] # 777번 아이템과 유사한 15개의 아이템 추천 리스트

사용자의 구매 기록(purchase)를 활용한 ALS모델 기반 추천 시스템

df_implicit = df[['user_id', 'product_id', 'event_type']]
df_implicit

df_implicit.info()

def event_type_score(df, value, score):
    df.loc[df['event_type'] == value, 'event_type'] = score

dict = {'view':np.nan, 'cart':np.nan, 'purchase':1}
for k, v in dict.items():
    event_type_score(df_implicit, k, v)

df_implicit = df_implicit.dropna()
df_implicit.shape

df_implicit

df_implicit['event_type'] = df_implicit['event_type'].astype('int')
df_implicit.info()

df_implicit_grouped = df_implicit.groupby(['product_id', 'user_id']).sum().reset_index()
df_implicit_grouped

user_unique = df_implicit_grouped['user_id'].unique()
product_unique = df_implicit_grouped['product_id'].unique()

user_to_idx = {v:k for k,v in enumerate(user_unique)}
product_to_idx = {v:k for k,v in enumerate(product_unique)}

temp_user_data = df_implicit_grouped['user_id'].map(user_to_idx.get).dropna()
if len(temp_user_data) == len(df_implicit_grouped):   # 모든 row가 정상적으로 인덱싱되었다면
    print('user_id column indexing OK!!')
    df_implicit_grouped['user_id'] = temp_user_data   # 데이터프레임['user_id']을 인덱싱된 Series로 교체해 줍니다. 
else:
    print('user_id column indexing Fail!!')

# product_to_idx을 통해 product_id 컬럼도 동일한 방식으로 인덱싱해 줍니다. 
temp_product_data = df_implicit_grouped['product_id'].map(product_to_idx.get).dropna()
if len(temp_product_data) == len(df_implicit_grouped):
    print('product column indexing OK!!')
    df_implicit_grouped['product_id'] = temp_product_data
else:
    print('product column indexing Fail!!')

df_implicit_grouped

num_user = df_implicit_grouped['user_id'].nunique()
num_product = df_implicit_grouped['product_id'].nunique()

implicit_matrix = sparse.csr_matrix((df_implicit_grouped.event_type, (df_implicit_grouped.user_id, df_implicit_grouped.product_id)),
                                    shape=(num_user,num_product))

implicit_matrix.T

matrix_size = implicit_matrix.shape[0]* implicit_matrix.shape[1]
num_purchases = len(implicit_matrix.nonzero()[0])
sparsity = 100 * (1 - (num_purchases / matrix_size))
sparsity

def make_train (matrix, percentage = .2):
    '''
    -----------------------------------------------------
    설명
    유저-아이템 행렬 (matrix)에서 
    1. 0 이상의 값을 가지면 1의 값을 갖도록 binary하게 테스트 데이터를 만들고
    2. 훈련 데이터는 원본 행렬에서 percentage 비율만큼 0으로 바뀜
    
    -----------------------------------------------------
    반환
    training_set: 훈련 데이터에서 percentage 비율만큼 0으로 바뀐 행렬
    test_set:     원본 유저-아이템 행렬의 복사본
    user_inds:    훈련 데이터에서 0으로 바뀐 유저의 index
    '''
    test_set = implicit_matrix.copy()
    
    training_set = implicit_matrix.copy()
    nonzero_inds = training_set.nonzero()
    nonzero_pairs = list(zip(nonzero_inds[0], nonzero_inds[1]))
    
    random.seed(0)
    num_samples = int(np.ceil(percentage * len(nonzero_pairs)))
    samples = random.sample(nonzero_pairs, num_samples)
    
    user_inds = [index[0] for index in samples]
    item_inds = [index[1] for index in samples]
    
    training_set[user_inds, item_inds] = 0
    training_set.eliminate_zeros()
    
    return training_set, test_set, list(set(user_inds))

# 훈련, 테스트 데이터 생성
mat_train, mat_test, product_users_altered = make_train(implicit_matrix, 0.2)

als_model = ALS(factors=30,
                regularization=0.1,
                alpha = 40,
                iterations = 30,
                calculate_training_loss = True)
als_model.fit(mat_train)

implicit.evaluation.precision_at_k(als_model, mat_train, mat_test, K=15, show_progress=True)

# 330394번째 유저의 latent factors 확인
als_model.user_factors[330393]

# 20772번째 제품의 latent factors 확인
als_model.item_factors[20771]

# 구매한 아이템과 비슷한 아이템에 대한 추천

product_idx = product_to_idx[1004650]
similar_product = als_model.similar_items(product_idx, N=16)

idx_to_product = {v:k for k,v in product_to_idx.items()}
product_id_list = [idx_to_product[i] for i in similar_product[0]]

df_cb[df_cb['product_id'] == 1004650]

df_cb[df_cb['product_id'].isin(product_id_list)]

# ID 519261186 유저가 선호할 만한 아이템 추천
user = user_to_idx[519261186]
# recommend에서는 user*item CSR Matrix를 받음.
product_recommend = als_model.recommend(user, implicit_matrix[user], N=15)
product_recommend

product_id_list2 = [idx_to_product[i] for i in product_recommend[0]]
df_cb[df_cb['product_id'].isin(product_id_list2)]

# ID 1002532 아이템을 추천한 이유를 추론 가능
explain = als_model.explain(user, implicit_matrix, itemid=672)
explain_list = [(idx_to_product[i[0]], i[1]) for i in explain[1]]
explain_product_list = []
for i in range(0,7):
    explain_product_list.append(explain_list[i][0])

df_cb[df_cb['product_id'].isin(explain_product_list)]

# als 모델 저장
als_model.save('als.npz')
